EES-3135 - cleanup slack test notifications

+++ tests/robot-tests/tests/libs/slack.py
@@ -9,7 +9,7 @@ from bs4 import BeautifulSoup
 PATH = f'{os.getcwd()}{os.sep}test-results'
 
 
-def _generate_slack_attachments(env: str):
+def _generate_slack_attachments(env: str, suite: str):
     with open(f'{PATH}{os.sep}output.xml', 'rb') as report:
         contents = report.read()
 
@@ -18,6 +18,14 @@ def _generate_slack_attachments(env: str):
 
     failed_tests = int(test['fail'])
     passed_tests = int(test['pass'])
+
+    failed_tests_field = {},
+
+    if failed_tests > 0:
+        failed_tests_field = {
+            "title": "Failed tests",
+            "value": failed_tests
+        }
 
     return [
         {
@@ -32,36 +40,42 @@ def _generate_slack_attachments(env: str):
                     "value": env
                 },
                 {
+                    "title": "Suite",
+                    "value": suite.replace('tests/', '')
+                },
+                {
                     "title": "Total test cases",
                     "value": passed_tests + failed_tests
                 },
-                {
-                    "title": "Passed tests",
-                    "value": passed_tests
-                },
-                {
-                    "title": "Failed tests",
-                    "value": failed_tests if failed_tests else "0"
-                },
-                {
-                    "title": "Results",
-                    "value": "Failed" if failed_tests else "Passed"
-                },
-
+                failed_tests_field,
             ]
         }
     ]
 
 
-def send_slack_report(env: str):
-    attachments = _generate_slack_attachments(env)
+def _tests_failed():
+    with open(f'{PATH}{os.sep}output.xml', 'rb') as report:
+        contents = report.read()
+
+        soup = BeautifulSoup(contents, 'lxml')
+        test = soup.find('total').find('stat')
+
+        failed_tests = int(test['fail'])
+
+        if failed_tests > 0:
+            return True
+
+
+def send_slack_report(env: str, suite: str):
+
+    attachments = _generate_slack_attachments(env, suite)
     data = {"attachments": attachments}
 
     webhook_url = os.getenv('SLACK_TEST_REPORT_WEBHOOK_URL')
     slack_bot_token = os.getenv('SLACK_BOT_TOKEN')
 
-    assert webhook_url, print("SLACK_TEST_REPORT_WEBHOOK_URL env variable needs to bet set")
-    assert slack_bot_token, print("SLACK_BOT_TOKEN env variable needs to bet set")
+    assert webhook_url, print("SLACK_TEST_REPORT_WEBHOOK_URL env variable needs to be set")
+    assert slack_bot_token, print("SLACK_BOT_TOKEN env variable needs to be set")
 
     response = requests.post(
         url=webhook_url,
@@ -71,16 +85,17 @@ def send_slack_report(env: str):
 
     print("Sent UI test statistics to #build")
 
-    client = WebClient(token=slack_bot_token)
-    print('Sending UI test report to #build')
+    if _tests_failed():
+        client = WebClient(token=slack_bot_token)
 
-    shutil.make_archive('UI-test-report', 'zip', PATH)
-    try:
-        client.files_upload(
-            channels='#build',
-            file='UI-test-report.zip',
-            title='test-report.zip',
-        )
-    except SlackApiError as e:
-        print(f'Error uploading test report: {e}')
-    os.remove('UI-test-report.zip')
+        shutil.make_archive('UI-test-report', 'zip', PATH)
+        try:
+            client.files_upload(
+                channels='#build',
+                file='UI-test-report.zip',
+                title='test-report.zip',
+            )
+        except SlackApiError as e:
+            print(f'Error uploading test report: {e}')
+        os.remove('UI-test-report.zip')
+    print('Sent UI test report to #build')
